reduce redundant guide text, include examples inline (fixes #2830) (#2873)
The main goal of this change is to remove copy/paste text from the guides. Instead we move common text to include files, which can be parameterized as needed.

Common text between the quickstart and kubernetes guide has been centralized. There was some drift, so the main include file does some branching based on parameters set at include time.

I also reworked the all-in-one zone/proxy/service names to be more consistent. (This will require a coordinated push of the docs-site and all-in-one:latest)

I also moved some other kubernetes guide text into includes with an eye toward reuse in the marathon guide in a future change.

Other changes:
- move the examples directory underneath guides, which allows me to both include and link to example files.
- replace references to `raw.githubusercontent.com` with links to `docs.turbinelabs.io`; this will make things more portable in the future, and makes sure that the links are always correct at time of _deploy_ rather than time of _push_.
- fix some yaml files that had unnecessary stuff in them
- fix many bad links
- actually link the custom kube doc

fbshipit-source-id: 47da56d