[deploytool] better handling of long branch names (#3088)
Problem: very long branch names break deploytool; while the tag is properly truncated to 63 chars, the name of the deployment includes a prefix, which extends the identifier beyond 63 chars, breaking deploytool, but only after you've build your dockerfile.

This change fixes that by including the prefix when doing the math the tag. Both the prefix and the branch name may get truncated, with some heuristics to keep things as meaningful as possible.

Since the prefix of the deployment name is now needed when calculating the tag, all deployment files including stage/tag went from this:

```yaml
name: api-server-{{stage}}-{{tag}}
```

to this:
```yaml
name: {{appendStageAndTag "api-server"}}
```

`versiontool.VersionProperties.KubeDeploymentName` is replaced with `versiontool.VersionProperties.AppendTag`, which takes a prefix and appends a properly sized tag inferred from the version metadata, possibly truncating the prefix as well.

Additionally, there's a `versiontool.AppendTag` which takes a fixed tag which it appends to a (possibly truncated) prefix. This exists for the case when a tag is specified manually.

fbshipit-source-id: 53259dd